update poly the ES6 Array.find with Array.Find
